make build details and build compare breadcrumb consistent
